refactor cpn
